requirements.txt
streamlit
torch
torchvision
segmentation-models-pytorch
Pillow
numpy
tensorflow
huggingface_hub
